refactor: rename `oneof` into `anyof` to fit implementation
