+++ src/Illuminate/Translation/lang/en/validation.php
@@ -117,7 +117,7 @@ return [
     'not_in' => 'The selected :attribute is invalid.',
     'not_regex' => 'The :attribute field format is invalid.',
     'numeric' => 'The :attribute field must be a number.',
-    'oneof' => 'The :attribute field does not match any of the allowed rule sets.',
+    'anyof' => 'The :attribute field does not match any of the allowed rule sets.',
     'password' => [
         'letters' => 'The :attribute field must contain at least one letter.',
         'mixed' => 'The :attribute field must contain at least one uppercase and one lowercase letter.',
